requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
Flask-SQLAlchemy>=3.1,<4.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0

# ML stack
pandas>=2.0,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.4,<2.0
matplotlib>=3.8,<4.0

# Optional on Python versions where wheel/build is supported.
# scikit-surprise>=1.1.4,<2.0

# Phase 4+
# tensorflow>=2.16,<3.0
